requirements.txt
warcio==1.7.5
bs4==0.0.2
dotenv==0.9.9
openai==1.106.0
PyYAML==6.0.2
tiktoken==0.11.0
llama-cpp-python==0.3.16
